user-interface/requirements.txt
streamlit>=1.0.0
requests>=2.25.1
Pillow>=8.0.0
